Use smart pointers for streams and effects

+++ src/TAudioChannel.h
@@ -100,11 +100,11 @@ class TAudioChannel
         void SetStream(TAudioStreamPtr stream);
         TAudioStreamPtr GetStream();
 
-        int GetLeft()
+        long GetLeft()
         {
             return fLeftOut;
         }
-        int GetRight()
+        long GetRight()
         {
             return fRightOut;
         }
